requirements.txt
discord.py
python-dotenv
easy-pil
webcolors